Extract ItemView into a separate route
