+++ src/components/Loading/DotSpinner.css
@@ -1,0 +1,48 @@
+.spinner {
+  width: inherit;
+  height: inherit;
+  display: flex;
+  align-items: center;
+
+  & .bounce1 {
+    -webkit-animation-delay: -0.32s;
+    animation-delay: -0.32s;
+  }
+
+  & .bounce2 {
+    -webkit-animation-delay: -0.16s;
+    animation-delay: -0.16s;
+  }
+
+  & > div {
+    width: 33%;
+    height: 33%;
+    margin: 0 8%;
+    background-color: rgba(0, 0, 0, 0.62);
+    border-radius: 100%;
+    display: inline-block;
+    -webkit-animation: sk-bouncedelay 1.4s infinite ease-in-out both;
+    animation: sk-bouncedelay 1.4s infinite ease-in-out both;
+  }
+}
+
+@-webkit-keyframes sk-bouncedelay {
+  0%,
+  80%,
+  100% { -webkit-transform: scale(0); }
+  40% { -webkit-transform: scale(1); }
+}
+
+@keyframes sk-bouncedelay {
+  0%,
+  80%,
+  100% {
+    -webkit-transform: scale(0);
+    transform: scale(0);
+  }
+
+  40% {
+    -webkit-transform: scale(1);
+    transform: scale(1);
+  }
+}
