+++ src/components/Loading/index.js
@@ -1,0 +1,2 @@
+export { default as ViewLoading } from './ViewLoading';
+export { default as PaneLoading } from './PaneLoading';
