+++ src/components/index.js
@@ -4,3 +4,4 @@ export { default as InstanceFilters, instanceFilterRenderer } from './InstanceFi
 export { default as HoldingsRecordFilters, holdingsRecordFilterRenderer } from './HoldingsRecordFilters';
 export { default as ItemFilters, itemFilterRenderer } from './ItemFilters';
 export { default as InstancesList } from './InstancesList';
+export { PaneLoading, ViewLoading } from './Loading';
