+++ src/routes/index.js
@@ -1,2 +1,2 @@
 export { default as InstancesRoute } from './InstancesRoute';
-export default {};
+export { default as ItemRoute } from './ItemRoute';
