+++ src/routes/withData.js
@@ -133,7 +133,6 @@ const dataManifest = {
     },
     records: 'mtypes',
   },
-
   loanTypes: {
     type: 'okapi',
     path: 'loan-types',
@@ -154,11 +153,12 @@ const withData = WrappedComponent => class WithDataComponent extends React.Compo
     mutator: PropTypes.object.isRequired,
   };
 
-  isLoading = () => {
+  isLoading = (includeWrapped) => {
     const { resources } = this.props;
+    const manifest = includeWrapped ? WithDataComponent.manifest : dataManifest;
 
-    for (const key in dataManifest) {
-      if (dataManifest[key].type === 'okapi' &&
+    for (const key in manifest) {
+      if (manifest[key].type === 'okapi' &&
         !(resources[key] && resources[key].hasLoaded)) {
         return true;
       }
