+++ src/views/index.js
@@ -1,3 +1,2 @@
 export { default as InstancesView } from './InstancesView';
-
-export default {};
+export { default as ItemView } from './ItemView';
